+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- [TD-7074] Support for [search after pagination](https://www.elastic.co/guide/en/elasticsearch/reference/current/paginate-search-results.html#search-after)
+
 ## [7.3.0] 2025-03-10
 
 ### Added