+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 ### Added
 
-- [TD-7053] Filter widget for multi-language ordering
+- [TD-7053]: 
+  - Filter widget for multi-language ordering
+  - Add locales only for translatable widgets
 
 ## [7.3.0] 2025-03-10