+++ CHANGELOG.md
@@ -6,6 +6,12 @@
 
 - [TD-7074] Support for [search after pagination](https://www.elastic.co/guide/en/elasticsearch/reference/current/paginate-search-results.html#search-after)
 
+## [7.4.0] 2025-04-09
+
+### Changed
+
+- License
+
 ## [7.3.0] 2025-03-10
 
 ### Added